requirements.txt
flask>=3.0
rich>=13.0
opencv-python>=4.8
Pillow>=10.0
numpy>=1.24
openvino>=2024.0
